Move multiparameter to its own module
