+++ qcodes/parameters/__init__.py
@@ -72,11 +72,11 @@ from .array_parameter import ArrayParameter
 from .group_parameter import Group, GroupParameter
 from .grouped_parameter import DelegateGroup, DelegateGroupParameter, GroupedParameter
 from .multi_channel_instrument_parameter import MultiChannelInstrumentParameter
+from .multi_parameter import MultiParameter
 from .parameter import (
     CombinedParameter,
     DelegateParameter,
     ManualParameter,
-    MultiParameter,
     Parameter,
     ScaledParameter,
     combine,
